refactor(MenuDropdown): relocate component to UI folder, replace default button with custom SMenuButton, and pass onClose to children
